requirements.txt
# Web Framework
flask==3.0.0
django==4.2.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# HTTP Clients
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Security & Auth
cryptography==41.0.7
pyjwt==2.8.0
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Testing (dev dependencies)
pytest==7.4.3
pytest-cov==4.1.0

# ML Libraries
scikit-learn==1.3.2
tensorflow==2.15.0
